refactor firebase hooks into firebase.js
contain all firebase initialization and firebase object creation to
firebase.js

This object can be altered as needed.

exported const variables currently supported:
- googleAuthProvider: provider for authentication model to use
- auth: access to firebase authenication module
- firestore: access to the development database

TODO: create development, test, and production environments to
initialize the firebase service.
