+++ src/App.js
@@ -20,34 +20,36 @@ import { useCollectionData } from 'react-firebase-hooks/firestore';
 // Configuration is okay to be public, though can be factored out.
 // In fact, all setup for app can be factored out.
 // Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
+// import { initializeApp } from "firebase/app";
+// import { getAnalytics } from "firebase/analytics";
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyC8_TMAHAg-9Fzq4kIBDzf9_veCPdqUHGY",
-  authDomain: "irl-among-us-d5453.firebaseapp.com",
-  projectId: "irl-among-us-d5453",
-  storageBucket: "irl-among-us-d5453.appspot.com",
-  messagingSenderId: "299437319897",
-  appId: "1:299437319897:web:1a4aff6578a93b98cd40c8",
-  measurementId: "G-1SS9R7WKDG"
-};
+// const firebaseConfig = {
+//   apiKey: "AIzaSyC8_TMAHAg-9Fzq4kIBDzf9_veCPdqUHGY",
+//   authDomain: "irl-among-us-d5453.firebaseapp.com",
+//   projectId: "irl-among-us-d5453",
+//   storageBucket: "irl-among-us-d5453.appspot.com",
+//   messagingSenderId: "299437319897",
+//   appId: "1:299437319897:web:1a4aff6578a93b98cd40c8",
+//   measurementId: "G-1SS9R7WKDG"
+// };
 
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
+// // Initialize Firebase
+// const app = initializeApp(firebaseConfig);
+// const analytics = getAnalytics(app);
 
-// Initialize application
-firebase.initializeApp(firebaseConfig)
+// // Initialize application
+// firebase.initializeApp(firebaseConfig)
 
-// firebase objects
-const auth = firebase.auth();
-const firestore = firebase.firestore();
+// // firebase objects
+// const auth = firebase.auth();
+// const firestore = firebase.firestore();
+
+
 
 // const App = () =>{
 //   return (
@@ -61,6 +63,7 @@ const firestore = firebase.firestore();
 // }
 function App(){
   return(
+    
     <Router>
        <Routes>
           <Route path="/" element={<Pages />} />
