+++ src/pages/homepage.js
@@ -12,45 +12,11 @@ import { redirect } from "react-router-dom";
 import {Link} from "react-router-dom";
 import "./homepage.css";
 
-import firebase from 'firebase/compat/app';
-import 'firebase/compat/firestore';
-import 'firebase/compat/auth';
-
 // react hooks
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { useCollectionData } from 'react-firebase-hooks/firestore';
 
-// Configuration is okay to be public, though can be factored out.
-// In fact, all setup for app can be factored out.
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
-
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyC8_TMAHAg-9Fzq4kIBDzf9_veCPdqUHGY",
-  authDomain: "irl-among-us-d5453.firebaseapp.com",
-  projectId: "irl-among-us-d5453",
-  storageBucket: "irl-among-us-d5453.appspot.com",
-  messagingSenderId: "299437319897",
-  appId: "1:299437319897:web:1a4aff6578a93b98cd40c8",
-  measurementId: "G-1SS9R7WKDG"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-
-// Initialize application
-firebase.initializeApp(firebaseConfig)
-
-// firebase objects
-const auth = firebase.auth();
-const firestore = firebase.firestore();
+import {auth, googleAuthProvider} from "../firebase";
 
 function SignOut() {
   return auth.currentUser && (
@@ -71,15 +37,14 @@ function Pages(){
     const navigate = useNavigate();
 
     const signInWithGoogle = () => {
-        const provider = new firebase.auth.GoogleAuthProvider();
-        auth.signInWithPopup(provider)
+        auth.signInWithPopup(googleAuthProvider)
         .then(function(){
             console.log("REDIRECTED")
             navigate('/adminpage');
         });
     }
 
-    firebase.auth().onAuthStateChanged(function(user) {
+    auth.onAuthStateChanged(function(user) {
         if (user) {
           // User is signed in.
         } else {
